sui_pg_db
Struct Db

pub struct Db(/* private fields */);

Implementations

impl Db

pub async fn for_write(database_url: Url, config: DbArgs) -> Result<Self>

Construct a new DB connection pool talking to the database at database_url that supports write and reads. Instances of Db can be cloned to share access to the same pool.

pub async fn for_read(database_url: Url, config: DbArgs) -> Result<Self>

Construct a new DB connection pool talking to the database at database_url that defaults to read-only transactions. Instances of Db can be cloned to share access to the same pool.

pub async fn connect(&self) -> Result<Connection<'_>>

Retrieves a connection from the pool. Can fail with a timeout if a connection cannot be established before the DbArgs::connection_timeout has elapsed.

pub fn state(&self) -> State

Statistics about the connection pool

pub async fn run_migrations( &self, migrations: Option<&'static EmbeddedMigrations>, ) -> Result<Vec<MigrationVersion<'static>>>

Run migrations on the database. Use Dieselโs embed_migrations! macro to generate the migrations parameter for your indexer.